refactor(monitor): remove monitor pkg barrel files
