+++ packages/monitor/src/__tests__/monitor-state.test.ts
@@ -3,9 +3,9 @@ import { errAsync, okAsync } from "neverthrow";
 import type pm2 from "pm2";
 import { beforeEach, describe, expect, it, vi } from "vitest";
 import { AppManager } from "../core/app-manager.js";
-import type { MonitorPlugin } from "../core/monitor-plugin-driver.js";
 import LaunchpadMonitor from "../launchpad-monitor.js";
 import type { MonitorConfig } from "../monitor-config.js";
+import type { MonitorPlugin } from "../monitor-plugin.js";
 
 // Mock process.exit to prevent tests from actually exiting
 // @ts-expect-error - mockImplementation returns undefined
