--- packages/monitor/src/index.ts
@@ -1,11 +1,0 @@
-import LaunchpadMonitor from "./launchpad-monitor.js";
-
-export * from "./core/monitor-plugin-driver.js";
-// export * from './windows-api.js'; // Includes optional dependencies, so not exported here
-export * from "./launchpad-monitor.js";
-export * from "./monitor-commands.js";
-export * from "./monitor-config.js";
-export * from "./monitor-events.js";
-export * from "./monitor-state.js";
-
-export default LaunchpadMonitor;
