+++ packages/monitor/src/monitor-plugin.ts
@@ -6,7 +6,7 @@ import {
 	type PluginDriver,
 } from "@bluecadet/launchpad-utils/plugin-driver";
 import type pm2 from "pm2";
-import type LaunchpadMonitor from "../launchpad-monitor.js";
+import type LaunchpadMonitor from "./launchpad-monitor.js";
 
 type MonitorHookContext = {
 	/**
